+++ ndarray-gdal.js
@@ -62,7 +62,31 @@ fromGDALDataType[gdal.GDT_Float64] = Float64Array;
  * @param {ProgressCb} [options.progress_cb]
  * @returns {ndarray.NdArray<2>}
  */
-function readArray(opts) {
+
+/**
+ * Read the selection region into the given ndarray or a new ndarray, async version.
+ *
+ * x, y specify the origin of the selection region and width and height specify its size.
+ * [x, y] default to [0, 0], [width, height] default to the full size of the raster data.
+ * If an existing array if passed in data, it would be used keeping its current stride.
+ * If the array has a different size than the selection region, the data will be resampled.
+ * The resampling algorithm can be specified in resampling, otherwise GDAL's default one will be used.
+ * If no array is specified, a new array of [width, height] size with a default
+ * positive/positive row-major stride will be allocated.
+ *
+ * @method readArrayAsync
+ * @param {ArrayOptions} [options]
+ * @param {ndarray.NdArray<2>} [options.data]
+ * @param {number} [options.x]
+ * @param {number} [options.y]
+ * @param {number} [options.width]
+ * @param {number} [options.height]
+ * @param {string} [options.resampling]
+ * @param {ProgressCb} [options.progress_cb]
+ * @returns {Promise<ndarray.NdArray<2>>}
+ */
+
+const makeReadArray = (fn) => function readArray(opts) {
   let { data, x, y, width, height, resampling, progress_cb } = opts || {};
   if (!y) y = 0;
   if (!x) x = 0;
@@ -89,7 +113,7 @@ function readArray(opts) {
     throw new TypeError(`Type ${data.dtype} is not supported by GDAL`);
   }
 
-  this.read(x, y, width, height, data.data, {
+  const r = fn.call(this, x, y, width, height, data.data, {
     buffer_height: data.shape[0],
     buffer_width: data.shape[1],
     data_type: gdalType,
@@ -100,8 +124,11 @@ function readArray(opts) {
     progress_cb
   });
 
+  if (typeof r.then === 'function') {
+    return r.then(() => data);
+  }
   return data;
-}
+};
 
 /**
  * Write the selection region from the given ndarray.
@@ -120,7 +147,25 @@ function readArray(opts) {
  * @param {ProgressCb} [options.progress_cb]
  * @returns {void}
  */
-function writeArray(opts) {
+
+/**
+ * Write the selection region from the given ndarray, async version.
+ *
+ * x, y specify the origin of the selection region and width and height specify its size.
+ * [x, y] default to [0, 0], [width, height] default to the size of the array.
+ * Resampling when writing is not supported by GDAL.
+ *
+ * @method writeArrayAsync
+ * @param {ArrayOptions} options
+ * @param {ndarray.NdArray<2>} options.data
+ * @param {number} [options.x]
+ * @param {number} [options.y]
+ * @param {number} [options.width]
+ * @param {number} [options.height]
+ * @param {ProgressCb} [options.progress_cb]
+ * @returns {Promise<void>}
+ */
+const makeWriteArray = (fn) => function writeArray(opts) {
   let { data, x, y, width, height, resampling, progress_cb } = opts || {};
 
   if (!data || !data.stride) {
@@ -145,7 +190,7 @@ function writeArray(opts) {
     throw new TypeError(`Type ${data.dtype} is not supported by GDAL`);
   }
 
-  this.write(x, y, width, height, data.data, {
+  return fn.call(this, x, y, width, height, data.data, {
     buffer_height: data.shape[0],
     buffer_width: data.shape[1],
     data_type: gdalType,
@@ -154,9 +199,9 @@ function writeArray(opts) {
     offset: data.offset,
     progress_cb
   });
+};
 
-  return data;
-}
-
-gdal.RasterBandPixels.prototype.readArray = readArray;
-gdal.RasterBandPixels.prototype.writeArray = writeArray;
+gdal.RasterBandPixels.prototype.readArray = makeReadArray(gdal.RasterBandPixels.prototype.read);
+gdal.RasterBandPixels.prototype.writeArray = makeWriteArray(gdal.RasterBandPixels.prototype.write);
+gdal.RasterBandPixels.prototype.readArrayAsync = makeReadArray(gdal.RasterBandPixels.prototype.readAsync);
+gdal.RasterBandPixels.prototype.writeArrayAsync = makeWriteArray(gdal.RasterBandPixels.prototype.writeAsync);