+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Allow specifying a progress callback
  - Generate `index.d.ts` with yatag
  - Switch the unit testing to TS
+ - Add {Read|Write}ArrayAsync()
 
 ## [1.0.0] 2021-05-29